[Tolk] Type `coins`, function `ton("0.05")`
A new compile-time function is introduced.
It converts a constant floating-point string to nanotoncoins.
Example: `ton("0.05")` is equal to 50000000.
It requires a constant string; `ton(some_var)` is an error.

Type of its return value is `coins`, not `int`.
Like `intN`, it's backed by a general int at TVM level,
all arithmetics (except addition) degrade to int.
Type `coins` will be serialized as varint16 in the future.