Optimize layout of and simplify vm::DataCell (#1570)
* Add missing includes to td/utils/ThreadSafeCounter.h

This allows the file to be included first in other headers.

* Optimize layout of and simplify vm::DataCell

This is a complete rewrite of vm::DataCell initially undertaken to make
optimizer life easier when working with DataCell. In particular, the
commit makes references to child cells reside at the known offset from
`this` of the data cell, removes a bunch of unneeded complexity from the
class definition itself (like `Info` struct), and uses placement new to
allocate class with a trailer instead of relying on template trickery
from `CellWithStorage.h`. Additionally, DataCell verification is
rewritten from scratch and collected into a nice `CellChecker` class to
make it easier to follow through and reason about both for humans and
compilers.

All in all, this is 3.5% time speedup on the verification benchmark
(which is somewhat hard to attribute to a particular change) and average
5.5% memory consumption improvement (most likely coming from
better packing of the fields).

* Store variable-sized data in `char trailer_[]` in vm::PrunnedCell

This is similar to how it is done in DataCell.

* Remove now unused vm/cells/CellWithStorage.h

* Make DynamicBagOfCellsDb::validate_bucket_a agnostic to use_arena option

This prevents test failures if we actually start destructing (but not
freeing) arena-allocated cells.