Add more CMake options (#1715)
* Enable colored diagnostics

A well-known limitation of using ninja is that it requires the user to
pass -fcolor-diagnostics=always flag to the compiler to get colored
diagnostics. Starting from version 3.24, CMake can do this for us if we
set CMAKE_COLOR_DIAGNOSTICS flag to a truthful value.

* Add CMake option to disable GDB backtrace on fatal signal

Backtraces we produce are very verbose and somewhat useless. They also
are sometimes so long that they fill all of terminal scrollback and hide
the actual error.

* Add an off-by-default CMake option for using LLD as a linker

Using LLD gives a 3-fold link-time improvement on my machine.